Subgrammars refactored in mixin classes
